+++ src/lib/ngsi/Throttling.h
@@ -38,8 +38,8 @@
 */
 typedef struct Throttling
 {
-  std::string  string;
-  unsigned int seconds;
+  std::string   string;
+  long long     seconds;
 
   Throttling();
   void         set(std::string value);
@@ -47,7 +47,7 @@ typedef struct Throttling
   bool         isEmpty(void);
   std::string  render(Format format, std::string indent, bool comma);
   std::string  check(RequestType requestType, Format format, std::string indent, std::string predetectedError, int counter);
-  int          parse(void);
+  long long    parse(void);
   void         present(std::string indent);
 } Throttling;
